refactor(StringAddCalculator): domain package로 이동
