+++ src/test/java/calculator/domain/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import org.junit.jupiter.api.Test;
